DOCS-89. Update the REST user guide with consistent wording and examples.
Updated the sample table layout in the KIJI schema user guide (to use a
formatted entity_id instead of the deprecated hash type) so that if a user
were to create that users table by copy/pasting the layout, and then followed
the REST examples, they would work. Also re-organized the layout of the pages
and cleaned up up some of the text and formatting so that it's consistent
with each other.

Reviewers: carol, shashir, milo